chore: fix issue in scheduling jobs queue & separate the job polling logic
