+++ packages/server/worker/src/index.ts
@@ -3,3 +3,4 @@ export { engineRunner } from './lib/engine'
 export * from './lib/executors/flow-job-executor'
 export * from './lib/utils/webhook-utils'
 export * from './lib/flow-worker'
+export * from './lib/job-polling'
