+++ packages/server/worker/src/lib/flow-worker.ts
@@ -1,25 +1,17 @@
 import { exceptionHandler, JobData, JobStatus, OneTimeJobData, QueueName, rejectedPromiseHandler, RepeatingJobData, system, WebhookJobData, WorkerSystemProps } from '@activepieces/server-shared'
 import { isNil } from '@activepieces/shared'
-import { Semaphore } from 'async-mutex'
 import { engineApiService, workerApiService } from './api/server-api.service'
 import { flowJobExecutor } from './executors/flow-job-executor'
 import { repeatingJobExecutor } from './executors/repeating-job-executor'
 import { webhookExecutor } from './executors/webhook-job-executor'
+import { pollJob } from './job-polling'
 
-const WORKER_CONCURRENCY = system.getNumberOrThrow(WorkerSystemProps.FLOW_WORKER_CONCURRENCY)
-const POLLING_CONCURRENCY = system.getNumberOrThrow(WorkerSystemProps.POLLING_CONCURRENCY)
+const FLOW_WORKER_CONCURRENCY = system.getNumberOrThrow(WorkerSystemProps.FLOW_WORKER_CONCURRENCY)
 const SCHEDULED_WORKER_CONCURRENCY = system.getNumberOrThrow(WorkerSystemProps.SCHEDULED_WORKER_CONCURRENCY)
-const SCHEDULED_POLLING_COUNT = system.getNumberOrThrow(WorkerSystemProps.SCHEDULED_POLLING_COUNT)
 
 let closed = true
 let workerToken: string
 let heartbeatInterval: NodeJS.Timeout
-
-const pollLocks = {
-    [QueueName.ONE_TIME]: new Semaphore(POLLING_CONCURRENCY),
-    [QueueName.SCHEDULED]: new Semaphore(POLLING_CONCURRENCY),
-    [QueueName.WEBHOOK]: new Semaphore(POLLING_CONCURRENCY),
-}
 
 export const flowWorker = {
     async init(generatedToken: string): Promise<void> {
@@ -31,10 +23,9 @@ export const flowWorker = {
     },
     async start(): Promise<void> {
         for (const queueName of Object.values(QueueName)) {
-            const times = queueName === QueueName.SCHEDULED ? SCHEDULED_WORKER_CONCURRENCY : WORKER_CONCURRENCY 
-            const pollingCount = queueName === QueueName.SCHEDULED ? SCHEDULED_POLLING_COUNT : 1
+            const times = queueName === QueueName.SCHEDULED ? SCHEDULED_WORKER_CONCURRENCY : FLOW_WORKER_CONCURRENCY 
             for (let i = 0; i < times; i++) {
-                rejectedPromiseHandler(run(queueName, pollingCount))
+                rejectedPromiseHandler(run(queueName))
             }
         }
     },
@@ -44,24 +35,18 @@ export const flowWorker = {
     },
 }
 
-async function run<T extends QueueName>(queueName: T, _pollingCount = 1): Promise<void> {
+async function run<T extends QueueName>(queueName: T): Promise<void> {
     while (!closed) {
         let engineToken: string | undefined
         try {
-            const jobs = await poll(workerToken, queueName, _pollingCount)
-
-            if (jobs.length === 0) {
+            const job = await pollJob(workerToken, queueName)
+            if (isNil(job)) {
                 continue
             }
-
-            const promises = jobs.map(async (job) => {
-                const { data, engineToken: jobEngineToken } = job
-                engineToken = jobEngineToken
-                await consumeJob(queueName, data, engineToken)
-                await markJobAsCompleted(queueName, engineToken)
-            })
-            
-            await Promise.all(promises)
+            const { data, engineToken: jobEngineToken } = job
+            engineToken = jobEngineToken
+            await consumeJob(queueName, data, engineToken)
+            await markJobAsCompleted(queueName, engineToken)
         }
         catch (e) {
             exceptionHandler.handle(e)
@@ -75,24 +60,6 @@ async function run<T extends QueueName>(queueName: T, _pollingCount = 1): Promis
                 )
             }
         }
-    }
-}
-
-async function poll(workerToken: string, queueName: QueueName, _pollingCount = 1) {
-    try {
-        await pollLocks[queueName].acquire(1)
-        const jobs = []
-        for (let i = 0; i < _pollingCount; i++) {
-            const job = await workerApiService(workerToken).poll(queueName)
-            if (isNil(job)) {
-                break
-            }
-            jobs.push(job)
-        }
-        return jobs
-    }
-    finally {
-        pollLocks[queueName].release(1)
     }
 }
